Switch from (Positive)IntegerField to BigIntegerField for the upload_length and upload_offset fields
The (Positive)IntegerField can only store integers up to 2147483647 (on Postgres at least), which would limit the upload size to about 2.1 GB.

+++ rest_framework_tus/models.py
@@ -25,8 +25,8 @@ class AbstractUpload(models.Model):
 
     state = FSMField(default=states.INITIAL)
 
-    upload_offset = models.PositiveIntegerField(default=0)
-    upload_length = models.IntegerField(default=-1)
+    upload_offset = models.BigIntegerField(default=0)
+    upload_length = models.BigIntegerField(default=-1)
 
     upload_metadata = JSONField(load_kwargs={'object_pairs_hook': collections.OrderedDict})
